re-factored and re-organized code to export `defineChatEndpoint` from root #70
