+++ src/tests/unit-tests/endpoint-chat-history.unit.test.ts
@@ -1,8 +1,8 @@
 import {
   defineChatEndpoint,
   getChatEndpointRunner,
-} from '../../endpoints/endpoints';
-import {setupGenkit} from '../../core/core';
+  setupGenkit,
+} from '../../index';
 import {InMemoryChatHistoryStore} from '../../history/in-memory-chat-history-store';
 
 /**
